chore(admin): rename admin components folder
