+++ client/components/about/AboutPage.js
@@ -4,8 +4,11 @@ class AboutPage extends React.Component {
   render() {
     return (
       <div className="card-panel z-depth-8" id="aboutcard">
-        <h5>Create, edit, and delete documents easily</h5>
-        <h5>Restrict access to your documents</h5>
+        <ul>
+          <h5> - Create, edit, and delete documents</h5>
+          <h5> - Restrict access to your documents</h5>
+          <h5> - Access Documents of persons in the same role as you</h5>
+        </ul>
       </div>
     );
   }
