+++ client/components/adminComponents/RoleList.js
@@ -12,10 +12,10 @@ const RoleList = ({ roles, editRole, deleteRole }) =>
                 <div className="col s4 offset s2 email">{role.title}</div>
                 <div className="col s2 name">{role.id}</div>
                 <div className="user-buttons row col s3">
-                  <a className="waves-effect waves-light btn"
+                  <a className="waves-effect waves-light btn blue-grey"
                     onClick={() => editRole(role)}>
                     <i className="tiny material-icons left">edit</i>edit</a>
-                  <a className="waves-effect waves-light btn"
+                  <a className="waves-effect waves-light btn blue-grey"
                     onClick={() => deleteRole(role.id)}>
                     <i className="tiny material-icons left">delete</i>delete</a>
                 </div>
